+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for variant record tags in semantic analysis.
+
 ## [0.25.0] - 2020-09-01
 
 ### Added